Add support for constant write mask
The write mask consumes 1 byte of flash and 1 byte of SRAM for every
register. This might be undesirable and/or unnecessary so allow the
write mask to be fixed with the preprocessor define
I2C_GLOBAL_WRITE_MASK. When this is defined, its value is always used,
and the implementation need not define i2c_w_mask.