Fix SDA open-collector support
For whatever reason, I gave up on getting this to work before. Now I
need 3v3 <-> 5V interop without a level shifter, so sort out
open-collector support.

Basically, the top bit of USIDR and the SDA pin bit in PORTB are ORed
together, I think that's what I didn't realise before.

With this change, PORTB's bit is always set high, and USIDR is filled
with '1' or '0' as appropriate for each stage of the transaction.

This should now mean that a high voltage is never driven onto the i2c
pins by the AVR.